remove tailwind
