+++ wc/.storybook/preview.js
@@ -1,6 +1,8 @@
 import { setCustomElementsManifestWithOptions } from './cemOpts';
 import manifest from '../custom-elements.json';
 
+import '../src/dsa-ui.css';
+
 // We are using the dist-custom-elements output target, so
 // we need to import any components we use separately.
 // Yes, it is "unused". Yes, it is necessary.
