+++ wc/package.json
@@ -35,7 +35,7 @@
     "generate": "stencil generate",
     "storybook": "start-storybook -p 6006",
     "build-storybook": "build-storybook",
-    "lint": "prettier --write {src,stories}/**/*{.ts,.tsx} && eslint {src,stories}/**/*{.ts,.tsx}",
+    "lint": "prettier --write {src,stories}/**/*{.ts,.tsx,.css} && eslint {src,stories}/**/*{.ts,.tsx}",
     "cem:build": "cem analyze --globs \"src/**/*.tsx\" --stencil",
     "bump-to": "yarn version --no-commit-hooks --no-git-tag-version --new-version",
     "pub": "yarn publish --access public"
@@ -69,10 +69,8 @@
     "postcss": "^8.4.17",
     "prettier": "2.8.0",
     "puppeteer": "^19.2.2",
-    "stencil-tailwind-plugin": "^1.6.0",
     "storybook-addon-designs": "^6.3.1",
     "storybook-dark-mode": "^1.1.2",
-    "tailwindcss": "^3.1.8",
     "typescript": "^4.8.4"
   },
   "license": "MIT"
