+++ wc/src/components/demo/my-component.css
@@ -1,0 +1,33 @@
+:host {
+  display: block;
+}
+
+.primary {
+  display: flex;
+  justify-content: center;
+  border-radius: 0.375rem;
+  --tw-bg-opacity: 1;
+  background-color: rgb(231 1 9);
+  padding: 1.5rem;
+  font-family: klima, ui-sans-serif, system-ui, -apple-system,
+    BlinkMacSystemFont, 'Segoe UI', Roboto, 'Helvetica Neue', Arial, 'Noto Sans',
+    sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji', 'Segoe UI Symbol',
+    'Noto Color Emoji';
+  --tw-text-opacity: 1;
+  color: rgb(255 255 255);
+}
+
+.secondary {
+  display: flex;
+  justify-content: center;
+  border-radius: 0.375rem;
+  --tw-bg-opacity: 1;
+  background-color: rgb(36 31 32);
+  padding: 1.5rem;
+  font-family: klima, ui-sans-serif, system-ui, -apple-system,
+    BlinkMacSystemFont, 'Segoe UI', Roboto, 'Helvetica Neue', Arial, 'Noto Sans',
+    sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji', 'Segoe UI Symbol',
+    'Noto Color Emoji';
+  --tw-text-opacity: 1;
+  color: rgb(255 255 255);
+}
