+++ wc/src/components/demo/test/my-component.spec.ts
@@ -1,0 +1,36 @@
+import { newSpecPage } from '@stencil/core/testing';
+import { MyComponent } from '../my-component';
+
+describe('my-component', () => {
+  it('renders', async () => {
+    const { root } = await newSpecPage({
+      components: [MyComponent],
+      html: '<my-component></my-component>',
+    });
+    expect(root).toEqualHtml(`
+      <my-component>
+        <mock:shadow-root>
+          <div class="primary">
+            Hello, World! I'm
+          </div>
+        </mock:shadow-root>
+      </my-component>
+    `);
+  });
+
+  it('renders with values', async () => {
+    const { root } = await newSpecPage({
+      components: [MyComponent],
+      html: `<my-component first="Stencil" last="'Don't call me a framework' JS"></my-component>`,
+    });
+    expect(root).toEqualHtml(`
+      <my-component first="Stencil" last="'Don't call me a framework' JS">
+        <mock:shadow-root>
+          <div class="primary">
+            Hello, World! I'm Stencil 'Don't call me a framework' JS
+          </div>
+        </mock:shadow-root>
+      </my-component>
+    `);
+  });
+});
