+++ wc/src/utils/utils.ts
@@ -1,4 +1,4 @@
-/* eslint-disable @typescript-eslint/no-var-requires */
+import globalStyles from '../global/dsa-ui.css';
 
 /*** Shortcut to readable class lists; not dynamic */
 export const formatClasses = (...args: string[]) => args.join(' ');
@@ -7,22 +7,20 @@ export const formatClasses = (...args: string[]) => args.join(' ');
  * Workaround for long-lived bug in Chrome that prevents font load from within
  * shadow dom https://bugs.chromium.org/p/chromium/issues/detail?id=336876
  * https://github.com/ionic-team/stencil/issues/2072
+ * Doubles as way to add global styles to dom
  */
-export const maybeLoadFonts = () => {
-  const id = 'dsa-ui-font-loader';
-  // const fontPath = './assets/fonts/klima';
+export const maybeLoadGlobal = () => {
+  const id = 'dsa-ui-styles-loader';
 
   // Add custom font to page DOM since font-face doesn't work within Shadow DOM.
   const element = document.querySelector(`style[id="${id}"]`);
 
   // Only inject the element if it's not yet present
   if (!element) {
-    // const css = generateCssFromDir(fontPath);
-
-    // Turns out that the polyfills for fs don't work and haven't been
-    // maintained. Hardcoding the css generated from the functions below.
-    const css = `@font-face{font-family:'klima';src:url('./assets/fonts/klima/klima-latin-200-italic.woff')format('woff'),url('./assets/fonts/klima/klima-latin-200-italic.woff2')format('woff2');font-weight:200;font-style:italic;}@font-face{font-family:'klima';src:url('./assets/fonts/klima/klima-latin-200-normal.woff')format('woff'),url('./assets/fonts/klima/klima-latin-200-normal.woff2')format('woff2');font-weight:200;font-style:normal;}@font-face{font-family:'klima';src:url('./assets/fonts/klima/klima-latin-300-italic.woff')format('woff'),url('./assets/fonts/klima/klima-latin-300-italic.woff2')format('woff2');font-weight:300;font-style:italic;}@font-face{font-family:'klima';src:url('./assets/fonts/klima/klima-latin-300-normal.woff')format('woff'),url('./assets/fonts/klima/klima-latin-300-normal.woff2')format('woff2');font-weight:300;font-style:normal;}@font-face{font-family:'klima';src:url('./assets/fonts/klima/klima-latin-400-italic.woff')format('woff'),url('./assets/fonts/klima/klima-latin-400-italic.woff2')format('woff2');font-weight:400;font-style:italic;}@font-face{font-family:'klima';src:url('./assets/fonts/klima/klima-latin-400-normal.woff')format('woff'),url('./assets/fonts/klima/klima-latin-400-normal.woff2')format('woff2');font-weight:400;font-style:normal;}@font-face{font-family:'klima';src:url('./assets/fonts/klima/klima-latin-500-italic.woff')format('woff'),url('./assets/fonts/klima/klima-latin-500-italic.woff2')format('woff2');font-weight:500;font-style:italic;}@font-face{font-family:'klima';src:url('./assets/fonts/klima/klima-latin-500-normal.woff')format('woff'),url('./assets/fonts/klima/klima-latin-500-normal.woff2')format('woff2');font-weight:500;font-style:normal;}@font-face{font-family:'klima';src:url('./assets/fonts/klima/klima-latin-700-italic.woff')format('woff'),url('./assets/fonts/klima/klima-latin-700-italic.woff2')format('woff2');font-weight:700;font-style:italic;}@font-face{font-family:'klima';src:url('./assets/fonts/klima/klima-latin-700-normal.woff')format('woff'),url('./assets/fonts/klima/klima-latin-700-normal.woff2')format('woff2');font-weight:700;font-style:normal;}@font-face{font-family:'klima';src:url('./assets/fonts/klima/klima-latin-900-italic.woff')format('woff'),url('./assets/fonts/klima/klima-latin-900-italic.woff2')format('woff2');font-weight:900;font-style:italic;}@font-face{font-family:'klima';src:url('./assets/fonts/klima/klima-latin-900-normal.woff')format('woff'),url('./assets/fonts/klima/klima-latin-900-normal.woff2')format('woff2');font-weight:900;font-style:normal;}@font-face{font-family:'klima';src:url('./assets/fonts/klima/klima-latin-950-italic.woff')format('woff'),url('./assets/fonts/klima/klima-latin-950-italic.woff2')format('woff2');font-weight:950;font-style:italic;}@font-face{font-family:'klima';src:url('./assets/fonts/klima/klima-latin-950-normal.woff')format('woff'),url('./assets/fonts/klima/klima-latin-950-normal.woff2')format('woff2');font-weight:950;font-style:normal;}`
-
+    // css file minified
+    const css = removeWhitespace(removeComments(globalStyles));
+    console.log(globalStyles);
+    console.log(css);
     const head = document.head || document.getElementsByTagName('head')[0];
     const style = document.createElement('style');
     style.appendChild(document.createTextNode(css));
@@ -65,9 +63,9 @@ export const maybeLoadFonts = () => {
 //     filenames.map((filename) => { const ext = filename.split('.').pop();
 //     return `url('${fontPath}/${filename}')format('${extFmtMap[ext]}')`
 //         });
-//         fontFace += `src:` + srcArr.join(`,`) + `;` 
-//           + `font-weight:${weight};` 
-//           + `font-style:${style};` 
+//         fontFace += `src:` + srcArr.join(`,`) + `;`
+//           + `font-weight:${weight};`
+//           + `font-style:${style};`
 //           + `}`;
 //         css += fontFace;
 //       }
@@ -80,3 +78,7 @@ export const maybeLoadFonts = () => {
 /*** Just in case we add more font formats! */
 // const extFmtMap = { woff: 'woff', woff2: 'woff2', ttf: 'truetype', otf:
 //   'opentype', eot: 'embedded-opentype'};
+
+const removeComments = (str: string) =>
+  str.replace(/\/\*[\s\S]*?\*\/|\/\/.*/g, '');
+const removeWhitespace = (str: string) => str.replace(/[\n\r\s]+/g, '');
