+++ wc/stories/my-component.stories.ts
@@ -5,6 +5,6 @@ export default {
   component: 'my-component',
 };
 
-const Template = (args) => `<my-component ${formatArgs(args)}></my-component>`;
+const Template = args => `<my-component ${formatArgs(args)}></my-component>`;
 
 export const Default = Template.bind({});
